+++ NEWS.md
@@ -1,5 +1,8 @@
 # readr 1.0.0.9000
 
+* `col_euro_double()`, `parse_euro_double()`, `col_numeric()`, and 
+  `parse_numeric()` have been removed.
+
 * `guess_encoding()` returns a tibble
 
 # readr 1.0.0